requirements.txt
flask==2.0.1
python-docx==0.8.11
Werkzeug==2.0.1
Pillow==8.4.0